Disable min, max, resize for dialogs and move User Settings, View Existing Sites, Delete Sites to Controls namespace (#281)
