+++ nvQuickSite/Controllers/FileSystemController.cs
@@ -18,12 +18,9 @@
 namespace nvQuickSite.Controllers
 {
     using System;
-    using System.Diagnostics;
     using System.IO;
     using System.Linq;
-    using System.Runtime.CompilerServices;
     using System.Security.AccessControl;
-    using System.Threading;
     using System.Xml.Linq;
 
     using nvQuickSite.Controllers.Exceptions;
