+++ nvQuickSite/Controllers/IISController.cs
@@ -18,7 +18,6 @@
 namespace nvQuickSite.Controllers
 {
     using System;
-    using System.Collections;
     using System.Collections.Generic;
     using System.Drawing;
     using System.Globalization;
