+++ nvQuickSite/Controls/Settings/UserSettings.resx
