+++ nvQuickSite/Controls/Sites/DeleteSiteProgress.resx
