+++ nvQuickSite/Controls/Sites/ViewExistingSites.resx
